Create prediction based on 3 models and inflating valuable customers' prediction
Create new columns in the test data for the predictions of the 3 models.
Create a new column 'operational_pred' in the test data which takes the adequate gold/silver or bronze model's prediction based on 'average cost min'
Create a 'final_pred' column:
For the top 40 observations sorted by 'average cost min' descending (so the most valuable customers) where the 'operational_pred' is higher than 0.5,
We overwrite the prediction to 1, 0.999999, 0.999998 etc.
This immitates that if we predict a customer to churn with 50%+ probability, and the customer has high 'average cost min', then we will try to retain that customer. We do
this by artificially inflating its prediction of churn. So even if a customer had barely higher "probability" than 0.5 but they had a large 'average cost min', we will
still give a high "probability" (close to 1) of churning. This will help elevate our TOP20 evaluation metric.

Create a submission dataset called xgboost_submission_3models.csv